+++ hedera-mirror-web3/src/test/java/com/hedera/services/evm/contracts/operations/HederaSelfDestructOperationV038Test.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hedera.mirror.web3.evm.contracts.operations;
+package com.hedera.services.evm.contracts.operations;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
@@ -23,8 +23,8 @@ import static org.mockito.BDDMockito.given;
 
 import com.hedera.mirror.web3.evm.store.contract.HederaEvmStackedWorldStateUpdater;
 import com.hedera.node.app.service.evm.contracts.operations.HederaExceptionalHaltReason;
-import com.hedera.services.evm.contracts.operations.HederaSelfDestructOperation;
 import java.util.function.BiPredicate;
+import java.util.function.Predicate;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.datatypes.Wei;
 import org.hyperledger.besu.evm.EVM;
@@ -39,11 +39,13 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-class HederaSelfDestructOperationTest {
+class HederaSelfDestructOperationV038Test {
 
-    private static final Address BENEFICIARY = Address.fromHexString("0x0000000000000000000000000000000000000929");
-    private static final String ETH_ADDRESS = "0xc257274276a4e539741ca11b590b9447b26a8051";
-    private static final Address EIP_1014_ETH_ADDRESS = Address.fromHexString(ETH_ADDRESS);
+    static final Address BENEFICIARY = Address.fromHexString("0x0000000000000000000000000000000000000929");
+    static final Address BENEFICIARY_SYSTEM_ACCOUNT =
+            Address.fromHexString("0x00000000000000000000000000000000000002ED");
+    static final String ETH_ADDRESS = "0xc257274276a4e539741ca11b590b9447b26a8051";
+    static final Address EIP_1014_ETH_ADDRESS = Address.fromHexString(ETH_ADDRESS);
 
     @Mock
     private HederaEvmStackedWorldStateUpdater worldUpdater;
@@ -63,11 +65,14 @@ class HederaSelfDestructOperationTest {
     @Mock
     private BiPredicate<Address, MessageFrame> addressValidator;
 
-    private HederaSelfDestructOperation subject;
+    @Mock
+    private Predicate<Address> systemAccountDetector;
+
+    private HederaSelfDestructOperationV038 subject;
 
     @BeforeEach
     void setUp() {
-        subject = new HederaSelfDestructOperation(gasCalculator, addressValidator);
+        subject = new HederaSelfDestructOperationV038(gasCalculator, addressValidator, systemAccountDetector);
 
         given(frame.getWorldUpdater()).willReturn(worldUpdater);
         given(gasCalculator.selfDestructOperationGasCost(any(), eq(Wei.ONE))).willReturn(2L);
@@ -110,12 +115,24 @@ class HederaSelfDestructOperationTest {
     void executeInvalidSolidityAddress() {
         givenRejectingValidator();
 
-        given(frame.getStackItem(0)).willReturn(BENEFICIARY);
+        given(frame.getStackItem(0)).willReturn(BENEFICIARY_SYSTEM_ACCOUNT);
 
         final var opResult = subject.execute(frame, evm);
 
         assertEquals(HederaExceptionalHaltReason.INVALID_SOLIDITY_ADDRESS, opResult.getHaltReason());
         assertEquals(2L, opResult.getGasCost());
+    }
+
+    @Test
+    void haltsWhenBeneficiaryIsSystemAccount() {
+        given(frame.getStackItem(0)).willReturn(BENEFICIARY);
+        given(frame.getRecipientAddress()).willReturn(EIP_1014_ETH_ADDRESS);
+        given(systemAccountDetector.test(any())).willReturn(true);
+        // when
+        final var result = subject.execute(frame, evm);
+        // then
+        assertEquals(HederaExceptionalHaltReason.INVALID_SOLIDITY_ADDRESS, result.getHaltReason());
+        assertEquals(2L, result.getGasCost());
     }
 
     private void givenRubberstampValidator() {
